More cleanup and fixing Typescript Errors
