--- src/components/Banner.tsx
@@ -1,7 +1,0 @@
-function Banner() {
-  return (
-    <div>Banner</div>
-  )
-}
-
-export default Banner
