+++ src/components/EmptyBox.tsx
@@ -1,0 +1,7 @@
+export function EmptyBox({children}: {children: any, style?: any, ref?: any, className: string}) {
+  return (
+    <div>
+      {children}
+    </div>
+  )
+}
