+++ src/components/Prose.tsx
@@ -1,6 +1,6 @@
 import clsx from 'clsx'
 
-export function Prose({ children, className }) {
+export function Prose({ children, className }: {children: any, className: string}) {
   return (
     <div className={clsx(className, 'prose dark:prose-invert')}>{children}</div>
   )
