+++ src/lib/generateRssFeed.tsx
@@ -8,13 +8,13 @@ export async function generateRssFeed() {
   let articles = await getAllArticles()
   let siteUrl = process.env.NEXT_PUBLIC_SITE_URL
   let author = {
-    name: 'Spencer Sharp',
-    email: 'spencer@planetaria.tech',
+    name: 'Trevor Campbell',
+    email: 'thatdude@trevordcampbell.com',
   }
 
   let feed = new Feed({
     title: author.name,
-    description: 'Your blog description',
+    description: `Trevor's little internet kingdom!`,
     author,
     id: siteUrl,
     link: siteUrl,
