+++ src/pages/_app.tsx
@@ -1,7 +1,7 @@
 import '../styles/globals.css'
 import { useEffect, useRef } from 'react'
 
-import { Header } from '@/components/HeaderOriginal'
+import { Header } from '@/components/Header'
 import { Footer } from '@/components/Footer'
 import 'focus-visible'
