+++ src/pages/articles/first-article.mdx
@@ -1,0 +1,28 @@
+import { ArticleLayout } from '@/components/ArticleLayout'
+
+export const meta = {
+  author: 'Trevor Campbell',
+  date: '2022-11-18',
+  title: `I'm sick in Jakarta ... anyway, here's my website!`,
+  description:
+    'When you’re building a website for a company as ambitious as Planetaria, you need to make an impression. I wanted people to visit our website and see animations that looked more realistic than reality itself.',
+}
+
+export default (props) => <ArticleLayout meta={meta} {...props} />
+
+If you're reading this, I'm probably already dead ☠️
+
+Just Kidding! While I may be sick in a foreign country nearly at the opposite point of the globe from home, lots of bed rest finally gave me the time to launch my personal website! I've been putting this off for a while, and travelling the globe was no help! But since getting sick in Jakarta, I decided to buckle down and finish it up!
+
+Some of you have not seen me in a few months – at the start of August, I packed my (enormous) bags and ventured off to Pakistan! After travelling the country for two whole months, I continued on to Singapore.
+
+## Why am I here?
+
+~~I ask myself this question each morning in panic.~~
+You're probably wondering what this website will be used for! This website gives me a place to post my writings, photographs, videos, and other fun things!
+
+Lots more to come, but for now, check out my photo gallery to see pictures from my travels in Asia!
+
+[Trevor's Photo Gallery](/gallery)
+
+See you soon! 👋
