+++ src/pages/articles/index.tsx
@@ -5,7 +5,7 @@ import { SimpleLayout } from '@/components/SimpleLayout'
 import { getAllArticles } from '@/lib/getAllArticles'
 import { formatDate } from '@/lib/formatDate'
 
-function Article({ article }) {
+function Article({ article }: {article: any}) {
   return (
     <article className="md:grid md:grid-cols-4 md:items-baseline">
       <Card className="md:col-span-3">
@@ -34,7 +34,7 @@ function Article({ article }) {
   )
 }
 
-export default function ArticlesIndex({ articles }) {
+export default function ArticlesIndex({ articles }: {articles: any}) {
   return (
     <>
       <Head>
@@ -50,7 +50,7 @@ export default function ArticlesIndex({ articles }) {
       >
         <div className="md:border-l md:border-zinc-100 md:pl-6 md:dark:border-zinc-700/40">
           <div className="flex max-w-3xl flex-col space-y-16">
-            {articles.map((article) => (
+            {articles.map((article: any) => (
               <Article key={article.slug} article={article} />
             ))}
           </div>
