+++ src/pages/uses.tsx
@@ -4,7 +4,7 @@ import { Card } from '@/components/Card'
 import { Section } from '@/components/Section'
 import { SimpleLayout } from '@/components/SimpleLayout'
 
-function ToolsSection({ children, ...props }) {
+function ToolsSection({ children, ...props }: {title?: string, children: any}) {
   return (
     <Section {...props}>
       <ul role="list" className="space-y-16">
@@ -14,7 +14,7 @@ function ToolsSection({ children, ...props }) {
   )
 }
 
-function Tool({ title, href, children }) {
+function Tool({ title, href, children }: {title: string, href?: string, children: any}) {
   return (
     <Card as="li">
       <Card.Title as="h3" href={href}>
